consolidate repository and kafka packages
