--- packages/kafka/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "dist"
-	},
-	"include": ["src/**/*.ts"],
-	"exclude": ["dist", "node_modules", "**/*.spec.ts"]
-}
