+++ packages/maestro-provider/src/kafka/processMessage/documentMessage.ts
