+++ packages/maestro-provider/src/kafka/processMessage/lyric/handleDocument.ts
