+++ packages/maestro-provider/src/kafka/processMessage/parser.ts
