+++ packages/maestro-provider/src/kafka/processMessage/song/handleDocument.ts
