--- packages/repository/.mocharc.json
@@ -1,7 +1,0 @@
-{
-	"extension": ["ts"],
-	"spec": "test/**/*.spec.ts",
-	"ignore": ["test/integration/**/*"],
-	"node-option": ["import=tsx"],
-	"timeout": 35000
-}
