--- packages/repository/.nycrc.json
@@ -1,7 +1,0 @@
-{
-	"check-coverage": true,
-	"all": true,
-	"include": ["src/**/*.ts"],
-	"reporter": ["html", "lcov", "text", "text-summary"],
-	"report-dir": "coverage"
-}
