--- packages/repository/src/index.ts
@@ -1,2 +1,0 @@
-export { repository } from './repositories';
-export { getRepoInformation } from './utils/utils';
